Move files into their corresponding folders
