+++ src/app/edit/components/Edit.tsx
@@ -2,17 +2,17 @@
 
 import { useState, useEffect, useRef } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import AppPage from '../apppage';
+import AppPage from '../../components/AppPage';
 import EditModal from './EditModal';
 import Undo from './Undo';
-import { AppTypes } from '../apps-manifest';
+import { AppTypes } from '../../scripts/apps';
 import {
 	hideApp,
 	addDeletedApp,
 	purgeApp,
 	purgeDeletedApp,
-} from './delete-app';
-import { TimeoutTypes, addTimeout, removeTimeout } from './timeout';
+} from '../scripts/delete';
+import { TimeoutTypes, addTimeout, removeTimeout } from '../scripts/timeout';
 
 export default function Edit() {
 	const [showModal, setShowModal] = useState(false);
