+++ src/app/edit/components/Undo.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { AppTypes } from '../apps-manifest';
+import { AppTypes } from '../../scripts/apps';
 import Image, { StaticImageData } from 'next/image';
 import { v4 as uuidv4 } from 'uuid';
