+++ src/app/edit/page.tsx
@@ -1,4 +1,4 @@
-import Edit from './Edit';
+import Edit from './components/Edit';
 
 export default function EditPage() {
 	return <Edit />;
