+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import AppPage from "./apppage";
+import AppPage from './components/AppPage';
 
 export default function Home() {
-  return <AppPage apps={[]} edit={false} />;
+	return <AppPage apps={[]} edit={false} />;
 }
